+++ packages/common/src/signHelpers.ts
@@ -3,36 +3,165 @@ import {
     AccountInfo,
     AccountTransaction,
     AccountTransactionSignature,
+    CredentialSignature,
+    HexString,
+    SimpleAccountKeys,
+    WalletExportFormat,
+    WithAccountKeys,
 } from './types';
 import * as ed from '@noble/ed25519';
 import { Buffer } from 'buffer/';
 import { AccountAddress } from './types/accountAddress';
 import { sha256 } from './hash';
+import { mapRecord } from './util';
 
+/**
+ * A structure to use for creating signatures on a given digest.
+ */
 export interface AccountSigner {
+    /**
+     * Creates a signature of the provided digest
+     *
+     * @param {Buffer} digest - The digest to create signatures on.
+     *
+     * @returns {Promise<AccountTransactionSignature>} A promise resolving with a set of signatures for a set of credentials corresponding to some account
+     */
     sign(digest: Buffer): Promise<AccountTransactionSignature>;
+    /**
+     * Returns the amount of signatures that the signer produces
+     */
     getSignatureCount(): bigint;
 }
 
+const getSignature = async (
+    digest: Buffer,
+    privateKey: HexString
+): Promise<HexString> =>
+    Buffer.from(await ed.sign(digest, privateKey)).toString('hex');
+
 /**
- * Creates a signer for an account which uses the first credential's first keypair.
+ * Creates an `AccountSigner` for an account which uses the first credential's first keypair.
  * Note that if the account has a threshold > 1 or the first credentials has a threshold > 1, the transaction signed using this will fail.
- * @param privateKey the ed25519 private key in HEX format. (First credential's first keypair's private key)
+ *
+ * @param {HexString} privateKey - the ed25519 private key in HEX format. (First credential's first keypair's private key)
+ *
+ * @returns {AccountSigner} an `AccountSigner` which creates a signature using the first credentials first keypair
  */
-export function buildBasicAccountSigner(privateKey: string): AccountSigner {
+export function buildBasicAccountSigner(privateKey: HexString): AccountSigner {
     return {
         getSignatureCount() {
             return 1n;
         },
         async sign(digest: Buffer) {
-            const signature = Buffer.from(
-                await ed.sign(digest, privateKey)
-            ).toString('hex');
             return {
                 0: {
-                    0: signature,
+                    0: await getSignature(digest, privateKey),
                 },
             };
+        },
+    };
+}
+
+const isWalletExport = <T extends WithAccountKeys>(
+    value: T | WalletExportFormat
+): value is WalletExportFormat =>
+    (value as WalletExportFormat).value?.accountKeys !== undefined;
+
+const isSimpleAccountKeys = <T extends WithAccountKeys>(
+    value: T | WalletExportFormat | SimpleAccountKeys
+): value is SimpleAccountKeys =>
+    (value as WalletExportFormat).value?.accountKeys === undefined &&
+    (value as T).accountKeys === undefined;
+
+const getKeys = <T extends WithAccountKeys>(
+    value: T | WalletExportFormat | SimpleAccountKeys
+): SimpleAccountKeys => {
+    if (isSimpleAccountKeys(value)) {
+        return value;
+    }
+    const { keys } = isWalletExport(value)
+        ? value.value.accountKeys
+        : value.accountKeys;
+
+    return mapRecord(keys, (credKeys) =>
+        mapRecord(credKeys.keys, (keyPair) => keyPair.signKey)
+    );
+};
+
+const getCredentialSignature = async (
+    digest: Buffer,
+    keys: Record<number, HexString>
+): Promise<CredentialSignature> => {
+    const sig: CredentialSignature = {};
+    for (const key in keys) {
+        sig[key] = await getSignature(digest, keys[key]);
+    }
+    return sig;
+};
+
+/**
+ * Creates an `AccountSigner` for an account exported from a Concordium wallet.
+ * Creating signatures using the `AccountSigner` will hold signatures for all credentials and all their respective keys included in the export.
+ *
+ * @param {WalletExportFormat} walletExport - The wallet export object.
+ *
+ * @returns {AccountSigner} An `AccountSigner` which creates signatures using all keys for all credentials
+ */
+export function buildAccountSigner(
+    walletExport: WalletExportFormat
+): AccountSigner;
+/**
+ * Creates an `AccountSigner` for an arbitrary format extending the {@link WithAccountKeys} type.
+ * Creating signatures using the `AccountSigner` will hold signatures for all credentials and all their respective keys included.
+ *
+ * @param {AccountKeys} value.accountKeys - Account keys of type {@link AccountKeys} to use for creating signatures
+ *
+ * @returns {AccountSigner} An `AccountSigner` which creates signatures using all keys for all credentials
+ */
+export function buildAccountSigner<T extends WithAccountKeys>(
+    value: T
+): AccountSigner;
+/**
+ * Creates an `AccountSigner` for the {@link SimpleAccountKeys} type.
+ * Creating signatures using the `AccountSigner` will hold signatures for all credentials and all their respective keys included.
+ *
+ * @param {SimpleAccountKeys} keys - Account keys to use for creating signatures
+ *
+ * @returns {AccountSigner} An `AccountSigner` which creates signatures using all keys for all credentials
+ */
+export function buildAccountSigner(keys: SimpleAccountKeys): AccountSigner;
+/**
+ * Creates an `AccountSigner` for an account which uses the first credential's first keypair.
+ * Note that if the account has a threshold > 1 or the first credentials has a threshold > 1, the transaction signed using this will fail.
+ *
+ * @param {HexString} key - The ed25519 private key in HEX format. (First credential's first keypair's private key)
+ *
+ * @returns {AccountSigner} An `AccountSigner` which creates a signature using the first credentials first keypair
+ */
+export function buildAccountSigner(key: HexString): AccountSigner;
+export function buildAccountSigner<T extends WithAccountKeys>(
+    value: T | WalletExportFormat | SimpleAccountKeys | string
+): AccountSigner {
+    if (typeof value === 'string') {
+        return buildBasicAccountSigner(value);
+    }
+
+    const keys = getKeys<T>(value);
+    const numKeys = Object.values(keys).reduce(
+        (acc, credKeys) => acc + BigInt(Object.keys(credKeys).length),
+        0n
+    );
+
+    return {
+        getSignatureCount() {
+            return numKeys;
+        },
+        async sign(digest: Buffer) {
+            const sig: AccountTransactionSignature = {};
+            for (const key in keys) {
+                sig[key] = await getCredentialSignature(digest, keys[key]);
+            }
+            return sig;
         },
     };
 }