+++ packages/common/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `CIS2Contract` class for interacting with smart contracts adhering to the CIS-2 standard.
 - `cis0Supports` function for checking standard support in smart contracts.
 - Made the `streamToList()` function public.
+- Build function `buildAccountSigner` for creating `AccountSigner` objects from genesis format, wallet export format, and a simple representation of credentials with keys.
 
 ## 6.4.2 2023-04-21